Change namespacing for work history components
