+++ app/components/candidate_interface/work_history_review_component.html.erb
