+++ app/components/work_history_component.html.erb
@@ -14,6 +14,6 @@
 
 <div data-qa="work-history">
   <% history.each do |item| %>
-    <%= render ProviderInterface::WorkHistoryItemComponent.new(item: item) %>
+    <%= render WorkHistoryItemComponent.new(item: item) %>
   <% end %>
 </div>
