+++ app/components/work_history_item_component.html.erb
