+++ app/views/provider_interface/application_choices/show.html.erb
@@ -44,7 +44,7 @@
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <%= render ProviderInterface::WorkHistoryComponent.new(application_form: @application_choice.application_form) %>
+    <%= render WorkHistoryComponent.new(application_form: @application_choice.application_form) %>
 
     <%= render ProviderInterface::VolunteeringHistoryComponent.new(application_form: @application_choice.application_form) %>
