+++ app/views/support_interface/application_forms/show.html.erb
@@ -37,8 +37,7 @@
 
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
-    <h2 class="govuk-heading-m govuk-!-font-size-27 govuk-!-margin-top-8">Work history</h2>
-    <%= render WorkHistoryReviewComponent.new(application_form: @application_form, editable: false, heading_level: 3) %>
+    <%= render WorkHistoryComponent.new(application_form: @application_form) %>
 
     <h2 class="govuk-heading-m govuk-!-font-size-27 govuk-!-margin-top-8">Unpaid experience and volunteering</h2>
     <%= render VolunteeringReviewComponent.new(application_form: @application_form, editable: false, heading_level: 3) %>
