+++ spec/components/previews/provider_interface/work_history_component_preview.rb
@@ -15,7 +15,7 @@ module ProviderInterface
 
     def render_component_for(application_form:)
       if application_form
-        render ProviderInterface::WorkHistoryComponent.new(application_form: application_form)
+        render WorkHistoryComponent.new(application_form: application_form)
       else
         render template: 'support_interface/docs/missing_test_data'
       end
